refactored types.AgentRelationshipRole into types.AgentRelationship
